+++ Fleece/Value.cc
@@ -148,9 +148,9 @@ namespace fleece {
             case kIntTag: {
                 int64_t i = asInt();
                 if (isUnsigned())
-                    sprintf(str, "%llu", (uint64_t)i);
+                    sprintf(str, "%llu", (unsigned long long)i);
                 else
-                    sprintf(str, "%lld", i);
+                    sprintf(str, "%lld", (long long)i);
                 break;
             }
             case kSpecialTag: {
